Functions and loops

+++ include/token.h
@@ -4,12 +4,12 @@
 #include <debug.h>
 
 typedef struct {
-    char*          buffer;
-    unsigned short buffer_size;
-    int            pos;
+    char*         buffer;
+    unsigned long buffer_size;
+    long          pos;
 } token_line_t;
 
-void set_tokenizer(char* buffer, unsigned short bsize, token_line_t* tl);
+void set_tokenizer(char* buffer, unsigned long bsize, token_line_t* tl);
 char get_next_token(token_line_t* tl);
 
 #endif
